requirements.txt
jinja2
SQLAlchemy~=2.0.27
openai~=0.27.10
uvicorn~=0.27.1
fastapi~=0.109.2
starlette~=0.36.3
python-dotenv~=1.0.1
python-multipart

# pip install fastapi
# pip install openai
# pip install uvicorn
# pip install sqlalchemy
# pip install starlette
# pip install python-dotenv
